Extend the default search pattern to support 'name:arch'
A search pattern with no explicit term, "s", is equivalent to
"?name(s)".  This changes modifies these semantics when the
pattern is arch-qualified, "name:arch" which is now equivalent
to "?exact-name(name) ?architecture(arch)".

These are the semantics used by apt-get and others which are
internally derived from pkgCache::FindPkg("name:arch").

On the command-line, it is possible to specify both an architecture
and an archive/version, as well as an override specifier:

 # aptitude install an:armel/sid tf:armel-

* src/generic/apt/matching/parse.cc:
  - extended semantics as per above.
* src/cmdline/cmdline_versions.cc:
  - modify argument processing to use the extended semantics.