Remove redundancy in doc build system
Lots of code was duplicated between makefiles for the
various translations.  Much of this is now in common
rules files making it easier to add new translations.

Translations that use po4a and provide full
documentation have been updated to use the new includes.
Some work is still required to make this compatible
with the non-po4a and partial translations.

* doc/common.mk:
  - common rules for doc makefiles
* doc/po4a.mk:
  - common rules for po4a makefiles
* doc/po4a/po4a.cfg:
  - new po files are picked up automatically
  - translate also safety-cost-level-diagram.svg
* doc/po4a/Makefile.am:
  - update-po only updates the po files
* doc/po4a/po/Makefile.am:
  - include aptitude.pot in dist files
* doc/{es,fr,it}/Makefile.am:
  - use common rules