Convert some instances of boost::shared_ptr and ::make_shared to std::
Convert some instances of boost::shared_ptr and boost::make_shared to the
corresponding versions now in the standard library.  The reason for this is
that it failed to compile straight away with the version in unstable
(boost-1.55) in some sections of the code (e.g. problemresolver), and
because since it is now part of the standard, there is little reason to use
Boost for that.  The rest of cases not covered in this version shall be
converted later on.