[curses] Use Yellow/Brown colour to highlight downgrades (Closes: #584510)
Previously downgrades were not highlighted, so it was bit confusing/odd and
it looked as if nothing would be done with the package.

There are not many colours "left" in the limited choice, only Yellow/Brown
(same as downloads) and Blue.  Since Blue is used in the menus, Yellow/Brown
is probably better; Downgrades and Downloads are never on the same screen.