Add missing APT_HAS_TRIGGERS handling, present since 2007
In file pkg_ver_item.cc, function pkg_ver_columnizer::setup_column(), the
switch/case handling for longstate was missing the options added in 2007 for
APT_HAS_TRIGGERS, present for stageflag case and both cases in
pkg_columnizer::setup_column() of pkg_columnizer.cc.

(In other words, it was missing in 1 of these 4 places where it is
necessary).