[curses] Avoid marking as auto/non-auto if already set (Closes: #265013)
It is unnecessary, but most importantly, checking if it's already set to the
desired value is much faster than calling libapt to do it.  For many years and
still at the moment, the call to libapt takes a few tenths of a second even in
fast machines, so it is a bit of a problem when working in slow machines and/or
marking many packages at a time.