Make images available in the french translation by copying them from the original document (Closes: #414150, #697736)
This is obviously suboptimal because the images are not translated, but
since 10 of the 11 were missing altogether, this is not making the problem
worse.  The remaining image of the copied ones was present in the french
translation, but was missing a key part of the interface described in the
surrounding text (it was still "wrong", and hopefully the new one is more
understandable in this context, even if in a different language).

The suboptimal solution is being taken at this point because the bugs have
been reported almost 9 years ago, and almost 2 years ago I requested help
from the Debian French l10n team, and there has not been any feedback so
far.  Perhaps having English images in the French translation (rather than
missing the images altogether) will spike somebody into action, to send us
a recent and localised capture to fix the problem properly ;-)