Abort with Failure rather than continue under many error conditions (Closes: #121313, #302784, #445034, #576212, #798320)
Abort with Failure when files fail to be fetched (Closes: #121313, #302784)

Abort with Failure (rather to continue with the installation) when packages fail
to be fetched -- except if APT::Get::Fix-Missing is present and set to true.
When this happens, the actions in the command line also return with non-zero
status.  (Closes: #121313, #302784)

[cmdline] Abort with Failure when actions cannot be taken
(Closes: #121313, #445034, #576212, #798320)

The Closes of #121313 overlap with other fixes, there are reports of each
case merged.