New util function: create_temporary_changelog_dir()
This functionality is similar to temp::file and temp::dir (and these were
used before), but apt v1.1 added a feature by which the process drops
privileges to _apt:nogroup when downloading files from the network, to not
have code running as root doing such tasks.

As a consequence, temp::dirs are owned by root and libapt cannot drop
privileges, causing unwanted annoying warnings (#806595).  Since aptitude
can run as root and normal users, it is not possible to let libapt create
files and copy them later: when running as normal users, it is not possible
to access those files to copy them.

This function supports other code to get around these problems.