Avoid extra code calling apt_load_cache() when already loaded
src/cmdline/cmdline_util.cc
src/ui_download_manager.cc

The cache just (re)loaded in last steps of download_manager (either
download_install_manager or download_update_manager).

Loading again is mostly harmless because the function when called detects
that it has already been loaded and quits, but duplicated code is not good
in general, and the call was modified several times lately, so it is always
a maintenance burden (not knowing / noticing that it was actually duplicated
code, otherwise would have removed it then).