Change default of ::RemoveObsoleteScore to 310 (from 600) before it's released
Otherwise the resolver is too eager to apply solutions removing packages (an
undesired solution that it's already much favoured in the last few years,
raising complaints) when local/obsolete are involved, even if a single upgrade
package would resolve the situation.

This affects only the case of local/obsolete packages, the favoured suggestions
to remove many packages rather than upgrading or keeping very few is still a
problem.